fix circular dependencies to use selectDataviewInstancesResolvedVisible
