+++ apps/fishing-map/features/debug/DebugMenu.tsx
@@ -6,7 +6,7 @@ import { selectLocationQuery } from 'routes/routes.selectors'
 import { useAppDispatch } from 'features/app/app.hooks'
 import { selectAllDatasets } from 'features/datasets/datasets.slice'
 import { debugDatasetsInDataviews, debugRelatedDatasets } from 'features/datasets/datasets.debug'
-import { selectAllDataviewInstancesResolved } from 'features/dataviews/selectors/dataviews.instances.selectors'
+import { selectAllDataviewInstancesResolved } from 'features/dataviews/selectors/dataviews.resolvers.selectors'
 import styles from './DebugMenu.module.css'
 import { DebugOption, selectDebugOptions, toggleOption } from './debug.slice'
