+++ apps/fishing-map/features/map/controls/MapControls.tsx
@@ -31,7 +31,7 @@ import { ROOT_DOM_ELEMENT } from 'data/config'
 import { useMapBounds } from 'features/map/map-bounds.hooks'
 import { selectIsGFWUser } from 'features/user/selectors/user.selectors'
 import { useMapErrorNotification } from 'features/map/overlays/error-notification/error-notification.hooks'
-import { selectDataviewInstancesResolved } from 'features/dataviews/selectors/dataviews.instances.selectors'
+import { selectDataviewInstancesResolved } from 'features/dataviews/selectors/dataviews.resolvers.selectors'
 import { useRootElement } from 'hooks/dom.hooks'
 import { isPrintSupported, MAP_IMAGE_DEBOUNCE } from '../MapScreenshot'
 import styles from './MapControls.module.css'
