+++ apps/fishing-map/features/map/popups/PopupByCategory.tsx
@@ -21,7 +21,7 @@ import { POPUP_CATEGORY_ORDER } from 'data/config'
 import { AsyncReducerStatus } from 'utils/async-slice'
 import { useMapViewport } from 'features/map/map-viewport.hooks'
 import { getDatasetTitleByDataview } from 'features/datasets/datasets.utils'
-import { selectAllDataviewInstancesResolved } from 'features/dataviews/selectors/dataviews.instances.selectors'
+import { selectAllDataviewInstancesResolved } from 'features/dataviews/selectors/dataviews.resolvers.selectors'
 import ComparisonRow from 'features/map/popups/categories/ComparisonRow'
 import WorkspacePointsTooltipSection from 'features/map/popups/categories/WorkspacePointsLayers'
 import DetectionsTooltipRow from 'features/map/popups/categories/DetectionsLayers'
