+++ apps/fishing-map/features/reports/activity/ReportActivityGraphSelector.tsx
@@ -8,7 +8,7 @@ import {
   REPORT_ACTIVITY_GRAPH_BEFORE_AFTER,
   REPORT_ACTIVITY_GRAPH_PERIOD_COMPARISON,
 } from 'data/config'
-import { selectActiveReportDataviews } from 'features/app/selectors/app.reports.selector'
+import { selectActiveReportDataviews } from 'features/dataviews/selectors/dataviews.selectors'
 import { useFitAreaInViewport } from 'features/reports/areas/reports.hooks'
 import { useSetReportTimeComparison } from 'features/reports/areas/reports-timecomparison.hooks'
 import { TrackCategory, trackEvent } from 'features/app/analytics.hooks'
