+++ apps/fishing-map/features/reports/vessel-groups/vessels/VesselGroupReportVesselsTable.tsx
@@ -6,7 +6,7 @@ import { IconButton } from '@globalfishingwatch/ui-components'
 import { EMPTY_FIELD_PLACEHOLDER } from 'utils/info'
 import { useLocationConnect } from 'routes/routes.hook'
 import { getDatasetsReportNotSupported } from 'features/datasets/datasets.utils'
-import { selectActiveReportDataviews } from 'features/app/selectors/app.reports.selector'
+import { selectActiveReportDataviews } from 'features/dataviews/selectors/dataviews.selectors'
 import { selectUserData } from 'features/user/selectors/user.selectors'
 import DatasetLabel from 'features/datasets/DatasetLabel'
 import { EMPTY_API_VALUES } from 'features/reports/areas/reports.config'
