+++ apps/fishing-map/features/resources/resources.hooks.ts
@@ -3,7 +3,7 @@ import { useSelector } from 'react-redux'
 import { Resource } from '@globalfishingwatch/api-types'
 import { useAppDispatch } from 'features/app/app.hooks'
 import { fetchResourceThunk } from 'features/resources/resources.slice'
-import { selectDataviewsResources } from 'features/dataviews/selectors/dataviews.instances.selectors'
+import { selectDataviewsResources } from 'features/dataviews/selectors/dataviews.resolvers.selectors'
 
 const useFetchResources = (resources: Resource[]) => {
   const dispatch = useAppDispatch()
