+++ apps/fishing-map/features/timebar/TimebarActivityGraph.tsx
@@ -6,7 +6,7 @@ import {
   HighlighterCallbackFn,
   HighlighterCallbackFnArgs,
 } from '@globalfishingwatch/timebar'
-import { selectActiveActivityDataviewsByVisualisation } from 'features/dataviews/selectors/dataviews.selectors'
+import { selectActiveActivityDataviewsByVisualisation } from 'features/timebar/timebar.selectors'
 import { useHeatmapActivityGraph } from 'features/timebar/TimebarActivityGraph.hooks'
 import { formatNumber } from 'utils/info'
 import { TimebarVisualisations } from 'types'
