+++ apps/fishing-map/features/workspace/legacy-activity-category.hook.ts
@@ -9,7 +9,7 @@ import {
   VIIRS_MATCH_DATAVIEW_SLUG,
 } from 'data/workspaces'
 import { useDataviewInstancesConnect } from 'features/workspace/workspace.hook'
-import { selectAllDataviewInstancesResolved } from 'features/dataviews/selectors/dataviews.instances.selectors'
+import { selectAllDataviewInstancesResolved } from 'features/dataviews/selectors/dataviews.resolvers.selectors'
 
 export const useHideLegacyActivityCategoryDataviews = () => {
   const actionDone = useRef(false)
