+++ apps/fishing-map/features/workspace/workspace.hook.ts
@@ -14,7 +14,7 @@ import {
   selectUrlViewport,
 } from 'routes/routes.selectors'
 import { useLocationConnect } from 'routes/routes.hook'
-import { selectDataviewInstancesResolved } from 'features/dataviews/selectors/dataviews.instances.selectors'
+import { selectDataviewInstancesResolved } from 'features/dataviews/selectors/dataviews.resolvers.selectors'
 import { useSetMapCoordinates } from 'features/map/map-viewport.hooks'
 import { useTimerangeConnect } from 'features/timebar/timebar.hooks'
 import { selectWorkspaceDataviewInstances } from './workspace.selectors'
